requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
pillow
pytesseract
pypdf
pymupdf
passporteye
python-dotenv
google-genai
langchain-google-genai
browser-use
playwright
